Viz: Factor the core Lir framework into its own package
