+++ ts-shared/l4-ladder-visualizer/package.json
@@ -72,6 +72,7 @@
   },
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
+    "@repo/layout-ir": "*",
     "@repo/viz-expr": "*",
     "@tailwindcss/cli": "^4.0.9",
     "@tailwindcss/vite": "^4.0.6",
