+++ ts-shared/l4-ladder-visualizer/src/lib/algebraic-graphs/dag.ts
@@ -1,4 +1,4 @@
-import type { Eq, Ord, HasId } from '$lib/utils.js'
+import type { Eq, Ord, HasId } from '@repo/layout-ir'
 import {
   DirectedAMGraph,
   makeDirectedConnectAdjacencyMap,
