+++ ts-shared/l4-ladder-visualizer/src/lib/displayers/flow/flow-props.ts
@@ -1,4 +1,4 @@
-import type { DisplayerProps } from '$lib/layout-ir/core.js'
+import type { DisplayerProps } from '@repo/layout-ir'
 import type { FunDeclLirNode } from '$lib/layout-ir/ladder-graph/ladder.svelte.js'
 import type { LadderEnv } from '$lib/ladder-env.js'
