+++ ts-shared/l4-ladder-visualizer/src/lib/displayers/flow/ladder-lir-to-sf.ts
@@ -1,4 +1,4 @@
-import type { LirContext, LirId } from '$lib/layout-ir/core.js'
+import type { LirContext, LirId } from '@repo/layout-ir'
 import type {
   LadderLirNode,
   LadderLirEdge,
