+++ ts-shared/l4-ladder-visualizer/src/lib/displayers/non-flow-props.ts
@@ -1,4 +1,4 @@
-import type { DisplayerProps } from '$lib/layout-ir/core.js'
+import type { DisplayerProps } from '@repo/layout-ir'
 import type { PathsListLirNode } from '$lib/layout-ir/node-paths-selection.js'
 import type { NNFLadderGraphLirNode } from '$lib/layout-ir/ladder-graph/ladder.svelte.js'
