+++ ts-shared/l4-ladder-visualizer/src/lib/displayers/paths-list.svelte
@@ -7,7 +7,7 @@
   } from '$lib/ui-primitives/toggle-group/index.js'
   import { useLadderEnv } from '$lib/ladder-env.js'
   import type { LinPathLirNode } from '$lib/layout-ir/ladder-graph/ladder.svelte.js'
-  import type { LirContext, LirId } from '$lib/layout-ir/core.js'
+  import type { LirContext, LirId } from '@repo/layout-ir'
 
   /************************
        Lir
