+++ ts-shared/l4-ladder-visualizer/src/lib/index.ts
@@ -4,9 +4,9 @@ export * from './data/viz-expr-to-lir.js'
 /** Ladder 'Env' */
 export * from './ladder-env.js'
 
-/** Lir nodes.
-This is an intermediate layer between the concrete UI and the abstract data (the Viz Expr types / schemas). */
-export * from './layout-ir/core.js'
+/** @repo/layout-ir contains the core Lir framework.
+The Lir nodes are basically an intermediate layer between the concrete UI and the abstract data (e.g. the Viz Expr types / schemas). */
+export * from '@repo/layout-ir'
 export * from './layout-ir/ladder-graph/ladder.svelte.js'
 export * from './layout-ir/node-paths-selection.js'
