+++ ts-shared/l4-ladder-visualizer/src/lib/layout-ir/ladder-graph/edge-attributes.ts
@@ -1,4 +1,4 @@
-import type { Eq } from '$lib/utils.js'
+import type { Eq } from '@repo/layout-ir'
 import 'core-js/actual/set'
 
 /********************************
