+++ ts-shared/l4-ladder-visualizer/src/lib/layout-ir/ladder-lir-source.ts
@@ -1,5 +1,5 @@
 import type { LadderEnv } from '$lib/ladder-env'
-import type { LirNodeInfo } from './core.js'
+import type { LirNodeInfo } from '@repo/layout-ir'
 
 export interface LadderLirSource<A, B> {
   toLir(nodeInfo: LirNodeInfo, env: LadderEnv, data: A): Promise<B>
