+++ ts-shared/layout-ir/eslint.config.js
@@ -1,0 +1,3 @@
+import { config } from '@repo/eslint-config/base'
+
+export default config
